Solve the fraudulent notification challenge for gh-3, what took me 4h.
An important lesson here is that the median using mid1+mid2/2 could make the math wrong because the result is not round. When dealing with integers is important to notice if there are some divisions that can make this happen. My first version did this and I needed to fix it by debugging the code.