Fast solution (2s), but slighly wrong (by some small percentage).
Who would say that manipulating vectors with erase and insert could be so slow compared to creating another vector and just copying values and then copying all the values back to the original vector? Well, it makes sense now that I said it.

This solution does that, but must have a bug, because in the input04.txt from the tests the results are slighly different from the gabarit:

| Our results  | Right results |
| -------------|---------------|
|    12186216  |      12309087 |
|  1255491686  |    1202639023 |
|   702439491  |     691828291 |
|  2100992997  |    2089430676 |
|   165224518  |     166937130 |

I think I must give up now and use ChatGPT to avoid to be too late for my technical interviews.