gh-3: Solve another Sherlock difficult task.
That was a very confused question for me because I misunderstood the
relation between counters of characters and frequencies of counters. The
use of maps to keep the same counters was the hard work, because the index
is the counter, but I need the frequency of the maximum counter to
the final if.

The lesson here is to never underestimate the problem, because there are
problems that just using more and more data structure is not going to
solve it. It need to clarify the ideas first and navigate into the
deeper issue.