gh-3: Solve reverse shuffle merge challenge (slow motion solution).
It really costs me more time to understand what the issue is about than to solve
it properly, but after I realized that I only needed to reverse the string and
search for all the permutations of the string until I find it everything became
clearer.

However, permutations costs a lot to long strings and that should be what is
failing 11 tests from 17. My next step.