gh-3: Solve reverse shuffle merge with long strings.
I was unable to implement the code even knowing my next step should be to
construct the string a traversing the string s and keeping the status of each
character. An indian video help me to solve this final question
https://youtu.be/_QQe2TQ2o_4
I did not see the code part, just the sketch part, but it was enough. What was
missing was the humbleness to get hands dirty and create some status machine and
see what happens. After some code and recode I could do it 'alone'.