gh-3: Solve swap nodes in binary tree (not the fast solution).
This took me some time to realize the intrinsics of using binary tress in the
array format. Besides, it was not a straighforward solution because there were
some minor bugs as the fact the input does not includes all the empty nodes. I
build it again using map, but this is not a very wise solution. Probably I need
for the fast solution to ponder a optimal strategy about this.