gh-3: Solve more tests from swap nodes (but not all).
This time I optimized the node build just filling the valid nodes, had some
issue about recording the depth in big test cases and solve it. Now I had a
problem similar with the depth and not sure how to solve it. Perhaps using the
node value instead of node number (because exponential is a bitch, and it is not
scaling so well with 500 values).