gh-3: Fixed all issues about swapping nodes from bin tree.
That was hard, but I made it by myself. I travelled around all the problems
about not having the depth recorded and not all children from every node and in
the 90 minutes of the game I realized I was complicating too much the solution,
that was just to map every node as a parent to a valid pair of branches. The
depth issue is resolved just adding one to the current depth.

Problem solve, lessons learned =).