gh-3: Add tripple sum challenge solution for small quantities.
This is a naive solution, but works with small quantities of numbers. The catch
in the next version is to get the limits of some ranges in the three sets,
perhaps ignoring for the moment the potential to multiply the range in all sets,
since there will be numbers that don't fit into some range.