gh-3: Solve machines problem with O(log n) solution.
The solution was searched in the discussion area. There was a description about
to find the right number of days by guessing in a binary search way. I
implemented like this and fix a overflow issue about getting the middle point of
a simple average. It is fast as hell now.