gh-3: Implement the dummy slow solution for max subarray sum issue.
I know that this won't solve the harder tests with lots of numbers. I just
wanted to clarify that there is a ingenuous solution before to figure out what
is the clever solution. We need to calculate the moduli of all the possible
combinations of subarrays before to know what is the maximum one?